+++ LeGO-LOAM/include/utility.h
@@ -44,16 +44,12 @@
 #include <thread>
 #include <mutex>
 
-#define PI 3.14159265
-
-using namespace std;
-
 typedef pcl::PointXYZI  PointType;
 
 typedef Eigen::Vector3f Vector3;
 
-extern const string pointCloudTopic = "/velodyne_points";
-extern const string imuTopic = "/imu/data";
+extern const std::string pointCloudTopic = "/velodyne_points";
+extern const std::string imuTopic = "/imu/data";
 
 // VLP-16
 extern const int N_SCAN = 16;
